+++ packages/terra-form-select/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 * Fixed
-  * Fixed selection did not work with space-key `form-single-select`.
-  * Fixed Enter did not work to expand the combo box example `form-single-select` .
+   * Fixed selection not working with space key in `form-single-select`.
+  * Fixed enter key not expanding the combo box in `form-single-select` .
 
 ## 6.46.0 - (August 25, 2023)